10/31/18 - Finished 2nd branch of 3x3x2 branching level!
Fixed bug where Wall turrets would fire in the wrong direction.
Added indicator that signals when enemies are hurt by a projectile
Added health pickups! Small health heals for 5 points, large for 10 points
Enemies have a 30% chance of dropping a small health pickup and a 20% chance of dropping a large health pickup

Finished the 2nd branch of the level, arguably the largest and hardest of the branches! We're almost there!
	Finishing the saber will be the most difficult part -_-. I'll save that for Friday if I have to.